requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
sqlalchemy==2.0.25
python-multipart==0.0.6
bcrypt==4.1.2
python-dotenv==1.0.0
httpx==0.26.0
itsdangerous==2.1.2